Added localizations. Moved ReasonOptions component into its own file
