+++ src/components/Section/History/Employment/ReasonLeft.test.jsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import { mount } from 'enzyme'
 import { i18n } from '../../../../config'
-import ReasonLeft, { ReasonOptions } from './ReasonLeft'
+import ReasonLeft from './ReasonLeft'
 
 describe('The reason left component', () => {
   it('no error on empty', () => {
@@ -84,21 +84,5 @@ describe('The reason left component', () => {
       component.find({type: 'radio'}).first().simulate('change')
       expect(updates).toBe(1)
     }
-  })
-
-  it('updates values', () => {
-    let updates = 0
-    const expected = {
-      name: 'peace_i_am_out',
-      Reason: 'Fired',
-      onUpdate: () => {
-        updates++
-      }
-    }
-    const component = mount(<ReasonOptions {...expected} />)
-    component.find('textarea').simulate('change')
-    component.find({type: 'text', name: 'month'}).simulate('change', { target: { value: '1' } })
-    component.find({type: 'radio', value: 'Quit'}).simulate('click')
-    expect(updates).toBe(3)
   })
 })
